LTX: Colossus works as a gov center for tribal/despotism
Tribal and despotism have much more trouble from distance to the palace. Some of the small wonders work as gov centers so that regular expansion is possible. This should slow down the expansion or allow the player to expand fast if it's of for the cities to have very little production.