LTX: bombers able to fire missiles+torpedoes. + bugfix
The bombers are able to fire missile units but only to adjacent tiles. This way the range of those units is not increased too much. The bombers are able to more missiles further and allowing the players to fire them from a plane would more than double the range. The ships can also have some missiles decreasing the risk of battles. However if the ship is killed the missile is most likely also lost since it will run out of fuel.